rename OrgAPI to OrgAPIClient + some clean up
