+++ tests/tests_integration/test_api/test_organization.py
@@ -1,17 +1,17 @@
 import pytest
 
 from cognite.client import CogniteClient
-from cognite.client._api.organization import OrgAPI
+from cognite.client._org_client import OrgAPIClient
 
 
 @pytest.fixture(scope="session")
-def org_api(cognite_client_cog_idp: CogniteClient) -> OrgAPI:
+def org_api(cognite_client_cog_idp: CogniteClient) -> OrgAPIClient:
     client = cognite_client_cog_idp
-    return OrgAPI(client.config, client._API_VERSION, client)
+    return OrgAPIClient(client.config, client._API_VERSION, client)
 
 
 class TestOrganizationAPI:
-    def test_get_organization(self, org_api: OrgAPI) -> None:
+    def test_get_organization(self, org_api: OrgAPIClient) -> None:
         organization = org_api._organization
         assert isinstance(organization, str)
         assert organization != ""
@@ -20,6 +20,6 @@ class TestOrganizationAPI:
         else:
             pytest.skip("Organization not supported")
 
-    def test_get_request_org_endpoint(self, org_api: OrgAPI) -> None:
+    def test_get_request_org_endpoint(self, org_api: OrgAPIClient) -> None:
         response = org_api._get("/principals")
         assert response.status_code == 200
