+++ tests/tests_unit/test_testing.py
@@ -1,8 +1,8 @@
 import pytest
 
 from cognite.client import ClientConfig, CogniteClient
-from cognite.client._api.organization import OrgAPI
 from cognite.client._api_client import APIClient
+from cognite.client._org_client import OrgAPIClient
 from cognite.client.credentials import Token
 from cognite.client.testing import CogniteClientMock, monkeypatch_cognite_client
 from tests.utils import all_mock_children, all_subclasses, get_api_class_by_attribute
@@ -11,8 +11,8 @@ from tests.utils import all_mock_children, all_subclasses, get_api_class_by_attr
 def test_ensure_all_apis_are_available_on_cognite_mock():
     mocked_apis = all_mock_children(CogniteClientMock())
     available = {v.__class__ for v in mocked_apis.values()}
-    # OrgAPI is a base API and should not mocked directly.
-    expected = set(all_subclasses(APIClient, exclude={OrgAPI}))
+    # OrgAPIClient is a base API and should not mocked directly.
+    expected = set(all_subclasses(APIClient, exclude={OrgAPIClient}))
     # Any new APIs that have not been added to CogniteClientMock?
     assert not expected.difference(available), f"Missing APIs: {expected.difference(available)}"
     # Any removed APIs that are still available on CogniteClientMock?
